Split device ID persistence utilities
